I think I cleared out the problem with updating ROIs
I added a notebook using the smaller dataset, which uses these changes
along with a different naming scheme for the cells so that it is
much easier to make certain the cells are being followed correctly
with respect to time.  Using it, I was able to manually
double-check the first 6 cells.

One thing I did not do: Jacques asked for a refresher on initialzing
the dictionaries; I got confused about which ones to write down and
so decided I would email him and ask before adding it.